+++ visualizer/src/Canvas/Tool/BrushCanvas.js
@@ -2,7 +2,7 @@ import { useSelector } from '@xstate/react';
 import React, { useEffect, useMemo, useRef } from 'react';
 
 import { useCanvas, useTool, useToolbar } from '../../ServiceContext';
-import { drawTrace, drawBrush } from '../imageUtils';
+import { drawTrace, drawBrush } from '../canvasUtils';
 
 const BrushCanvas = ({ className }) => {
   const canvas = useCanvas();
